+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Added
+ - `GetItFutureBuilder3` for three dependencies
+
 ## [0.2.0] - 2023-06-09
 
 ### Added